// src/components/Typography/internal/Typography.types.ts
export type ITypographyElement =
  | 'p'
  | 'h1'
  | 'h2'
  | 'h3'
  | 'h4'
  | 'code'
  | 'pre'
  | 'span'

export type ITypographyFont = 'inter' | 'poppins'

export type ITypographySize = 'small' | 'base' | 'h4' | 'h3' | 'h2' | 'h1'

export type ITypographyWeight = 'regular' | 'medium' | 'semi' | 'bold'